js3

+++ src/main/java/com/bbs/demo/repository/UsersRepository.java
@@ -7,7 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bbs.demo.model.Users;
 
-public interface UsersRepository extends JpaRepository<Users, Long> {
+public interface UsersRepository extends JpaRepository<Users, Integer> {
 	Optional<Users> findByUsername(String username);  
+	
 }
+
